% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalAssetCatelog.R
\name{DigitalAssetCatelog_view}
\alias{DigitalAssetCatelog_view}
\title{数据资产目录查询}
\usage{
DigitalAssetCatelog_view(erp_token, where_clause)
}
\arguments{
\item{where_clause}{}
}
\value{
两个数的和
}
\description{
数据资产目录查询
}
\examples{
DigitalAssetCatelog_view
}
