% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPsrcView.R
\name{LPsrcView_view}
\alias{LPsrcView_view}
\title{LinkProd的src表查询}
\usage{
LPsrcView_view(erp_token, FDataTableName)
}
\arguments{
\item{erp_token}{}
}
\value{
两个数的和
}
\description{
LinkProd的src表查询
}
\examples{
LPsrcView_view
}
